requirements.txt
appdirs==1.4.0
click==6.6
Flask==0.12
Flask-OAuth==0.12
gunicorn==19.7.0
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8.1
MarkupSafe==0.23
oauth2==1.9.0.post1
olefile==0.44
packaging==16.8
Pillow==4.0.0
pymongo==3.4.0
pyparsing==2.1.10
python-magic==0.4.12
six==1.10.0
Werkzeug==0.11.15
xlrd==1.0.0
